UFO Nested Secondary Index Bug Fix
- Support the case when a field marked as secondary index is not set (as it is optional) + test
- Move secondary index tests to its own file
- Add UFO Streaming API Unit Tests
