+++ test/src/test/resources/proto/sample_schema.proto
@@ -82,6 +82,22 @@ message SampleTableBMsg {
     optional string payload = 1;
 }
 
+message SampleTableCMsg {
+    option (org.corfudb.runtime.table_schema).requires_backup_support = true;
+    option (org.corfudb.runtime.table_schema).ownership_validation = true;
+    option (org.corfudb.runtime.table_schema).is_federated = true;
+
+    optional string payload = 1;
+}
+
+message SampleTableDMsg {
+    option (org.corfudb.runtime.table_schema).stream_tag = "sample_streamer_4";
+    option (org.corfudb.runtime.table_schema).requires_backup_support = false;
+    option (org.corfudb.runtime.table_schema).is_federated = true;
+
+    optional string payload = 1;
+}
+
 extend google.protobuf.MessageOptions {
     optional ManagedResourceOptionsMsg mgoptions = 54312;
 }
